Moved GSE JS files and lease app out of GSL repo for 0.8.5 release
GSE specific JS files have been moved to the goldstone-gse repository.  Additional
Tickets opened for moving python specific code to goldstone-gse (GOLD-256)
and for creating a relationship between the two codebases (GOLD-257). Petitioning
for close of this one as sufficient deliverable for 0.8.5 release.

Lease settings have been commented out.  Waffle is not available to check the flag
in the settings file, so we'll need to find another way to conditionally include
the GSE apps/settings.  This should be addressed with GOLD-257 in the 2.0
timeframe.

Ticket: GOLD-247