GOLD-266 Add API stats chart
This chart is pure d3 (not d3js) which should be more flexible
(for example, the shaded area between two lines) and should
set us up nicely for streaming or long-poll data consumption.

I had to hack on the legend, axes, and tooltip and pulled in a
couple of MIT licensed libraries to help, but got there.  There
is still room to improve the layout, and to start building our
own library of reusable tools.

Specific areas to improve: tooltip markers could be vertical
lines and "snap" to the center (maybe use rectangles instead
of circles); legend could be horizontal, legend could use
lines instead of circles.