GOLD-301, GOLD-305, GOLD-267 completed.
Info popovers on resource charts display correct dates, and autohide after 3 seconds.
The technique has an IE10 minimum, but there is a workaround function documented for
backward compatibility here:
https://developer.mozilla.org/en-US/docs/Web/API/Window.setTimeout#Callback_arguments

Topology hover icons do their best to stay on the screen, and are set to hide
when rolling off the icon, but also have a 3 second timer to hide them since some
slip past the event handler.