Convert Node model to Django backend from Elasticutils, bump project version
GOLD-608:
All model, serializer, view, test, and client code was adjusted to use the SQL
backend.  These changes extend to LoggingNode as well.  The model field
admin_disabled was changed to managed.  It was converted from a boolean field
to a char field due to the way django-filter handles booleans.  Truth was
reversed throughout the app to reflect the change in the field meaning.  The
model field last_seen_method was changed to update_method.  Additional logic
was added in process_host_stream to not save node when managed is false.

GOLD-593:
Version bumped to 2.0.1

Installer cleanup:
Added installation notes for ES and Logstash repos. Changed startup requirements
for celery to include httpd, elasticsearch, and redis. Changed startup order
for celery to come after httpd.