Ready for front-end review.
The data payload has changed a little bit.  The numbers are not reported in
fractions of a second.  That's ok, but will require the client to either
multiple by 1000 or change the axis label to be seconds.  There may be
additional data in the payload as well, but let's just replicate our existing
functionality for now.