EOD:   - removed logstash resubmission logic since upstream bug was addressed   - added agent logstash input for TCP 5516   - added filter to morph an agent fim event into a celery task body   - added debugging output for agent events
Next up will be to send the morphed event via the redis output and see
if celery can pick it up.  We used to do something like this via a custom
output, but I think we can do it with the vanilla redis output that is
now available.

We may also need to find a more elegant solution to handling logstash
conf and patterns from addons.  Currently everything is in the OSS
project. We shouldn't need any patterns, so not really leaking any
prop data other than the existence of fim.